util: all .c -> .cpp: converts to c++.
buffer.c -> buffer.cpp.

ecl_version, hash, hash_node, hash_sll -> CPP.

lookup_table, mzran, node_ctype, node_data -> CPP.

parser, path_stack, perm_vector, rng -> CPP.

set, statistics, string_util, stringlist -> CPP.

struct_vector, test_util, test_work_area, time_interval -> CPP.

timer, type_vector -> CPP

util -> CPP.

util_abort_gnu, util_abort_simple, util_endian, util_getuid -> CPP.

removed most warnings, c->cpp.

util_lfs, util_lockf, util_opendir, util_spawn -> CPP.

util_symlink, util_unlink, util_zlib, vectpor -> CPP.

vector_template -> CPP.

c -> CPP working in libres.

a few symbols moved to include.

util_copy_file__ removed from include.

util: new HPP files.

...

util.h: API changed back.

...

+++ lib/include/ert/util/util.h
@@ -270,7 +270,7 @@ typedef enum {left_pad   = 0,
   void         util_safe_free(void *);
   void         util_free_stringlist(char **, int );
   void         util_free_NULL_terminated_stringlist(char ** string_list);
-  char       * util_alloc_substring_copy(const char *, int offset , int N);
+  char       * util_alloc_substring_copy(const char *, size_t offset , int N);
   bool         util_is_directory(const char * );
   bool         util_is_file(const char * );
